urllist for yacy

+++ details.php
@@ -124,7 +124,7 @@ if ($username!="") {
 	if (($format=="website") or ($format=="information")) {
 		print "<a rel=\"nofollow\" target=\"_blank\" class=\"ym-button ym-play\" href=\"".$linkage."\">Website</a>\n";
 	}
-	if (($username==$owner) or ($username=="admin")) {
+	if ((($username==$owner) or ($username=="admin")) and ($dataset=="")) {
 		print "<a rel=\"nofollow\" href=\"edit.php?id=".$id."\" class=\"ym-button ym-edit\">Edit</a>";
 		print "<a rel=\"nofollow\" href=\"delete.php?id=".$id."\" class=\"ym-button ym-delete\">Delete</a>";
 		if ($format=="service") print "<a rel=\"nofollow\" href=\"add_service.php?url=".rawurlencode($wms)."\" class=\"ym-button ym-star\">Refresh</a>";
